Ensure known plugins are parsed from config
ESLint changed the access of `PLUGIN_NAME_PREFIX` constant -- it is no
longer exported. It is a well-know prefix, so we can just hard code it
on our side.

Since we didn't test for plugins in the config, we didn't catch this
issue. A test has been added as part of this commit.